[=] Fix input gathering
Input was requested from two different places (inside the interpreter
and in the main loop), but now it is only requested from the main loop.
The interpreter enters a WaitingInput state when it is waiting for an
input and the main loop can check and send its input to the interpreter;
this fixes the problem where you had to press multiple times to get one
input...

Numbers now send the number's value (0, 1, 2 ... 9) instead of their ascii
value (48, 49, 50 ... 57) because some programs require a zero input to
stop reading (so it is easier to be done now).